requirements.txt
anthropic
requests
schedule
Pillow
numpy
